refactor: Continue cleaning the code
