+++ package.json
@@ -14,14 +14,12 @@
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^4.21.2",
-    "multer": "^1.4.5-lts.2",
     "path": "^0.12.7",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
     "@types/cors": "^2.8.19",
     "@types/express": "^4.17.17",
-    "@types/multer": "^1.4.12",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6"
   }
